Move SPIClass definition
